fix norminette

+++ get_next_line.h
@@ -6,18 +6,18 @@
 /*   By: aassaf <aassaf@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/01 09:32:39 by aassaf            #+#    #+#             */
-/*   Updated: 2023/12/05 18:45:31 by aassaf           ###   ########.fr       */
+/*   Updated: 2023/12/06 13:38:25 by aassaf           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef GET_NEXT_LINE_H
 # define GET_NEXT_LINE_H
 
+# include <fcntl.h>
+# include <stdio.h>
 # include <stdlib.h>
 # include <string.h>
 # include <unistd.h>
-# include <fcntl.h>
-#include <stdio.h>
 
 # ifndef BUFFER_SIZE
 #  define BUFFER_SIZE 10
@@ -29,16 +29,16 @@ typedef struct s_list
 	struct s_list	*next;
 }					t_list;
 
-char *get_next_line(int fd);
-t_list	*ft_lstlast(t_list *lst);
-void        read_and_store(int fd, t_list **lst);
-void add_to_lst(t_list **lst, char *buff, int nb_read);
-int found_newline(t_list *lst);
-void create_line(char **line, t_list *lst);
-int	ft_strlen(const char *s);
-t_list	*ft_lstlast(t_list *lst);
-void copy_line(t_list *lst, char **line);
-void free_all(t_list *lst);
-void clear_lst(t_list **lst);
+char				*get_next_line(int fd);
+t_list				*ft_lstlast(t_list *lst);
+void				read_and_store(int fd, t_list **lst);
+void				add_to_lst(t_list **lst, char *buff, int nb_read);
+int					found_newline(t_list *lst);
+void				create_line(char **line, t_list *lst);
+size_t				ft_strlen(const char *s);
+t_list				*ft_lstlast(t_list *lst);
+void				copy_line(t_list *lst, char **line);
+void				free_all(t_list *lst);
+void				clear_lst(t_list **lst);
 
 #endif
